@@ -28,14 +28,14 @@ FLAGS = None
 
 def deepnn(x):
     """deepnn builds the graph for a deep net for classifying digits.
-  Args:
-  x: an input tensor with the dimensions (N_examples, 784), where 784 is the
-  number of pixels in a standard MNIST image.
-  Returns:
-  A tuple (y, keep_prob). y is a tensor of shape (N_examples, 10), with values
-  equal to the logits of classifying the digit into one of 10 classes (the
-  digits 0-9).
-  """
+    Args:
+    x: an input tensor with the dimensions (N_examples, 784), where 784 is the
+    number of pixels in a standard MNIST image.
+    Returns:
+    A tuple (y, keep_prob). y is a tensor of shape (N_examples, 10), with values
+    equal to the logits of classifying the digit into one of 10 classes (the
+    digits 0-9).
+    """
     # Reshape to use within a convolutional neural net.
     # Last dimension is for "features" - there is only one here, since images are
     # grayscale -- it would be 3 for an RGB image, 4 for RGBA, etc.
