@@ -1,18 +1,10 @@
 """Higher-level layer abstractions built on TF Encrypted."""
 from __future__ import absolute_import
 
-from tf_encrypted.keras import backend
 from tf_encrypted.keras import engine
 from tf_encrypted.keras import layers
 from tf_encrypted.keras import losses
 from tf_encrypted.keras import models
 from tf_encrypted.keras.models import Sequential
 
-__all__ = [
-    "backend",
-    "engine",
-    "layers",
-    "losses",
-    "models",
-    "Sequential",
-]
+__all__ = ["backend", "engine", "layers", "losses", "models", "Sequential"]
