@@ -5,7 +5,6 @@ from tf_encrypted.keras.engine.input_layer import Input
 from tf_encrypted.keras.layers.activation import Activation
 from tf_encrypted.keras.layers.convolutional import Conv2D
 from tf_encrypted.keras.layers.convolutional import DepthwiseConv2D
-from tf_encrypted.keras.layers.core import Reshape
 from tf_encrypted.keras.layers.dense import Dense
 from tf_encrypted.keras.layers.flatten import Flatten
 from tf_encrypted.keras.layers.normalization import BatchNormalization
@@ -14,6 +13,7 @@ from tf_encrypted.keras.layers.pooling import GlobalAveragePooling2D
 from tf_encrypted.keras.layers.pooling import GlobalMaxPooling2D
 from tf_encrypted.keras.layers.pooling import MaxPooling2D
 from tf_encrypted.keras.layers.relu import ReLU
+from tf_encrypted.keras.layers.reshape import Reshape
 
 __all__ = [
     "Input",
