@@ -5,16 +5,16 @@ from tf_encrypted.keras.engine import Layer
 
 class Activation(Layer):
     """Applies an activation function to an output.
-  Arguments:
-      activation: name of activation function to use or
-          TF Encrypted operation.
-  Input shape:
-      Arbitrary. Use the keyword argument `input_shape`
-      (tuple of integers, does not include the samples axis)
-      when using this layer as the first layer in a model.
-  Output shape:
-      Same shape as input.
-  """
+    Arguments:
+        activation: name of activation function to use or
+            TF Encrypted operation.
+    Input shape:
+        Arbitrary. Use the keyword argument `input_shape`
+        (tuple of integers, does not include the samples axis)
+        when using this layer as the first layer in a model.
+    Output shape:
+        Same shape as input.
+    """
 
     def __init__(self, activation, **kwargs):
         super(Activation, self).__init__(**kwargs)
