@@ -12,18 +12,19 @@ from tf_encrypted.keras.testing_utils import layer_test
 
 @pytest.mark.layers
 class TestActivation(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_activation_relu(self):
         self._core_activation(activation="relu")
 
     def _core_activation(self, **layer_kwargs):
         agreement_test(
-            tfe.keras.layers.Activation, kwargs=layer_kwargs, input_shape=[1, 5],
+            tfe.keras.layers.Activation,
+            kwargs=layer_kwargs,
+            batch_input_shape=[1, 5],
         )
         layer_test(
-            tfe.keras.layers.Activation, kwargs=layer_kwargs, batch_input_shape=[1, 5],
+            tfe.keras.layers.Activation,
+            kwargs=layer_kwargs,
+            batch_input_shape=[1, 5],
         )
 
     def test_backward_sigmoid(self) -> None:
@@ -43,34 +44,27 @@ class TestActivation(unittest.TestCase):
             loss = dense_out_pond * w
 
             # backward
-            d_out = w
-            _, d_x = tfe_layer.backward(d_out)
+            _, d_x = tfe_layer.backward(w)
 
-            with tfe.Session() as sess:
-                sess.run(tf.global_variables_initializer())
-                tfe_loss = sess.run(loss.reveal())
-                tfe_d_x = sess.run(d_x.reveal())
+            tfe_loss = loss.reveal().to_native()
+            tfe_d_x = d_x.reveal().to_native()
 
-        # reset graph
-        tf.reset_default_graph()
-
-        with tf.Session() as sess:
+        with tf.name_scope("TF"):
 
             tf_layer = tf.keras.layers.Activation("sigmoid", input_shape=[4])
 
             x = tf.Variable(input_data, dtype=tf.float32)
-            y = tf_layer(x)
             w = tf.Variable(weights_second_layer, dtype=tf.float32)
-            loss = y * w
 
-            # backward
-            d_x = tf.gradients(xs=x, ys=loss)
+            with tf.GradientTape() as tape:
+                y = tf_layer(x)
+                loss = y * w
 
-            sess.run(tf.global_variables_initializer())
-            tf_loss, tf_d_x = sess.run([loss, d_x])
+                # backward
+                d_x = tape.gradient(loss, x)
 
-            np.testing.assert_array_almost_equal(tfe_loss, tf_loss, decimal=1)
-            np.testing.assert_array_almost_equal(tfe_d_x, tf_d_x[0], decimal=2)
+            np.testing.assert_array_almost_equal(tfe_loss, loss, decimal=1)
+            np.testing.assert_array_almost_equal(tfe_d_x, d_x, decimal=2)
 
 
 if __name__ == "__main__":
