@@ -8,81 +8,81 @@ from tensorflow.python.keras.utils import conv_utils
 
 import tf_encrypted as tfe
 from tf_encrypted.keras import activations
-from tf_encrypted.keras import backend as KE
 from tf_encrypted.keras.engine import Layer
 from tf_encrypted.keras.layers.layers_utils import default_args_check
 from tf_encrypted.protocol import TFEPrivateTensor
+from tf_encrypted.protocol import TFEPrivateVariable
 
 logger = logging.getLogger("tf_encrypted")
 
 
 class Conv2D(Layer):
     """2D convolution layer (e.g. spatial convolution over images).
-  This layer creates a convolution kernel that is convolved
-  with the layer input to produce a tensor of
-  outputs. If `use_bias` is True,
-  a bias vector is created and added to the outputs. Finally, if
-  `activation` is not `None`, it is applied to the outputs as well.
-  When using this layer as the first layer in a model,
-  provide the keyword argument `input_shape`
-  (tuple of integers, does not include the sample axis),
-  e.g. `input_shape=(128, 128, 3)` for 128x128 RGB pictures
-  in `data_format="channels_last"`.
-  Arguments:
-      filters: Integer, the dimensionality of the output space
-          (i.e. the number of output filters in the convolution).
-      kernel_size: An integer or tuple/list of 2 integers, specifying the
-          height and width of the 2D convolution window.
-          Can be a single integer to specify the same value for
-          all spatial dimensions.
-      strides: An integer or tuple/list of 2 integers,
-          specifying the strides of the convolution along the height and width.
-          Can be a single integer to specify the same value for
-          all spatial dimensions.
-          Specifying any stride value != 1 is incompatible with specifying
-          any `dilation_rate` value != 1.
-      padding: one of `"valid"` or `"same"` (case-insensitive).
-      data_format: A string,
-          one of `channels_last` (default) or `channels_first`.
-          The ordering of the dimensions in the inputs.
-          `channels_last` corresponds to inputs with shape
-          `(batch, height, width, channels)` while `channels_first`
-          corresponds to inputs with shape
-          `(batch, channels, height, width)`.
-          It defaults to the `image_data_format` value found in your
-          Keras config file at `~/.keras/keras.json`.
-          If you never set it, then it will be "channels_last".
-      dilation_rate: an integer or tuple/list of 2 integers, specifying
-          the dilation rate to use for dilated convolution.
-          Can be a single integer to specify the same value for
-          all spatial dimensions.
-          Currently, specifying any `dilation_rate` value != 1 is
-          incompatible with specifying any stride value != 1.
-      activation: Activation function to use.
-          If you don't specify anything, no activation is applied
-          (ie. "linear" activation: `a(x) = x`).
-      use_bias: Boolean, whether the layer uses a bias vector.
-      kernel_initializer: Initializer for the `kernel` weights matrix.
-      bias_initializer: Initializer for the bias vector.
-      kernel_regularizer: Regularizer function applied to
-          the `kernel` weights matrix.
-      bias_regularizer: Regularizer function applied to the bias vector.
-      activity_regularizer: Regularizer function applied to
-          the output of the layer (its "activation")..
-      kernel_constraint: Constraint function applied to the kernel matrix.
-      bias_constraint: Constraint function applied to the bias vector.
-  Input shape:
-      4D tensor with shape:
-      `(samples, channels, rows, cols)` if data_format='channels_first'
-      or 4D tensor with shape:
-      `(samples, rows, cols, channels)` if data_format='channels_last'.
-  Output shape:
-      4D tensor with shape:
-      `(samples, filters, new_rows, new_cols)` if data_format='channels_first'
-      or 4D tensor with shape:
-      `(samples, new_rows, new_cols, filters)` if data_format='channels_last'.
-      `rows` and `cols` values might have changed due to padding.
-  """
+    This layer creates a convolution kernel that is convolved
+    with the layer input to produce a tensor of
+    outputs. If `use_bias` is True,
+    a bias vector is created and added to the outputs. Finally, if
+    `activation` is not `None`, it is applied to the outputs as well.
+    When using this layer as the first layer in a model,
+    provide the keyword argument `input_shape`
+    (tuple of integers, does not include the sample axis),
+    e.g. `input_shape=(128, 128, 3)` for 128x128 RGB pictures
+    in `data_format="channels_last"`.
+    Arguments:
+        filters: Integer, the dimensionality of the output space
+            (i.e. the number of output filters in the convolution).
+        kernel_size: An integer or tuple/list of 2 integers, specifying the
+            height and width of the 2D convolution window.
+            Can be a single integer to specify the same value for
+            all spatial dimensions.
+        strides: An integer or tuple/list of 2 integers,
+            specifying the strides of the convolution along the height and width.
+            Can be a single integer to specify the same value for
+            all spatial dimensions.
+            Specifying any stride value != 1 is incompatible with specifying
+            any `dilation_rate` value != 1.
+        padding: one of `"valid"` or `"same"` (case-insensitive).
+        data_format: A string,
+            one of `channels_last` (default) or `channels_first`.
+            The ordering of the dimensions in the inputs.
+            `channels_last` corresponds to inputs with shape
+            `(batch, height, width, channels)` while `channels_first`
+            corresponds to inputs with shape
+            `(batch, channels, height, width)`.
+            It defaults to the `image_data_format` value found in your
+            Keras config file at `~/.keras/keras.json`.
+            If you never set it, then it will be "channels_last".
+        dilation_rate: an integer or tuple/list of 2 integers, specifying
+            the dilation rate to use for dilated convolution.
+            Can be a single integer to specify the same value for
+            all spatial dimensions.
+            Currently, specifying any `dilation_rate` value != 1 is
+            incompatible with specifying any stride value != 1.
+        activation: Activation function to use.
+            If you don't specify anything, no activation is applied
+            (ie. "linear" activation: `a(x) = x`).
+        use_bias: Boolean, whether the layer uses a bias vector.
+        kernel_initializer: Initializer for the `kernel` weights matrix.
+        bias_initializer: Initializer for the bias vector.
+        kernel_regularizer: Regularizer function applied to
+            the `kernel` weights matrix.
+        bias_regularizer: Regularizer function applied to the bias vector.
+        activity_regularizer: Regularizer function applied to
+            the output of the layer (its "activation")..
+        kernel_constraint: Constraint function applied to the kernel matrix.
+        bias_constraint: Constraint function applied to the bias vector.
+    Input shape:
+        4D tensor with shape:
+        `(samples, channels, rows, cols)` if data_format='channels_first'
+        or 4D tensor with shape:
+        `(samples, rows, cols, channels)` if data_format='channels_last'.
+    Output shape:
+        4D tensor with shape:
+        `(samples, filters, new_rows, new_cols)` if data_format='channels_first'
+        or 4D tensor with shape:
+        `(samples, new_rows, new_cols, filters)` if data_format='channels_last'.
+        `rows` and `cols` values might have changed due to padding.
+    """
 
     def __init__(
         self,
@@ -92,6 +92,7 @@ class Conv2D(Layer):
         padding="valid",
         data_format=None,
         dilation_rate=(1, 1),
+        groups=1,
         activation=None,
         use_bias=True,
         kernel_initializer="glorot_uniform",
@@ -133,6 +134,7 @@ class Conv2D(Layer):
 
         # Not implemented arguments
         default_args_check(dilation_rate, "dilation_rate", "Conv2D")
+        default_args_check(groups, "groups", "Conv2D")
         default_args_check(kernel_regularizer, "kernel_regularizer", "Conv2D")
         default_args_check(bias_regularizer, "bias_regularizer", "Conv2D")
         default_args_check(activity_regularizer, "activity_regularizer", "Conv2D")
@@ -172,10 +174,12 @@ class Conv2D(Layer):
         if self.data_format != "channels_first":
             inputs = tfe.transpose(inputs, perm=[0, 3, 1, 2])
 
-        outputs = tfe.conv2d(inputs, self.kernel, self.strides[0], self.padding)
+        outputs = tfe.conv2d(
+            inputs, self.kernel.read_value(), self.strides, self.padding
+        )
 
         if self.use_bias:
-            outputs = outputs + self.bias
+            outputs = outputs + self.bias.read_value()
 
         if self.data_format != "channels_first":
             outputs = tfe.transpose(outputs, perm=[0, 2, 3, 1])
@@ -189,7 +193,7 @@ class Conv2D(Layer):
     def backward(self, d_y):
         x = self._layer_input
         y = self._layer_output
-        kernel = self.weights[0]
+        kernel = self.weights[0].read_value()
         grad_weights = []
 
         # Convert to NCHW format
@@ -224,7 +228,7 @@ class Conv2D(Layer):
 
         # Back prop for dx:
         # https://medium.com/@mayank.utexas/backpropagation-for-convolution-with-strides-8137e4fc2710
-        d_x = tfe.conv2d(padded_d_y, flipped_kernel, 1, "VALID")
+        d_x = tfe.conv2d(padded_d_y, flipped_kernel, [1, 1], "VALID")
 
         # Remove padding, if any
         if self.padding == "SAME":
@@ -247,7 +251,7 @@ class Conv2D(Layer):
         # Back prop for dw:
         # https://medium.com/@mayank.utexas/backpropagation-for-convolution-with-strides-fb2f2efc4faa
         # Output is in IOHW format
-        d_kernel = tfe.conv2d(x, inner_padded_d_y, 1, "VALID")
+        d_kernel = tfe.conv2d(x, inner_padded_d_y, [1, 1], "VALID")
         # Convert to HWIO
         d_kernel = tfe.transpose(d_kernel, perm=[2, 3, 0, 1])
         if self.lazy_normalization:
@@ -292,9 +296,9 @@ class Conv2D(Layer):
         h_filter, w_filter, _, n_filters = self.kernel_shape
 
         if self.data_format == "channels_first":
-            n_x, _, h_x, w_x = input_shape.as_list()
+            n_x, _, h_x, w_x = input_shape
         else:
-            n_x, h_x, w_x, _ = input_shape.as_list()
+            n_x, h_x, w_x, _ = input_shape
 
         if self.padding == "SAME":
             h_out = int(np.ceil(float(h_x) / float(self.strides[0])))
@@ -312,66 +316,66 @@ class Conv2D(Layer):
 class DepthwiseConv2D(Conv2D):
     """Depthwise separable 2D convolution.
 
-  Depthwise Separable convolutions consists in performing
-  just the first step in a depthwise spatial convolution
-  (which acts on each input channel separately).
-  The `depth_multiplier` argument controls how many
-  output channels are generated per input channel in the depthwise step.
+    Depthwise Separable convolutions consists in performing
+    just the first step in a depthwise spatial convolution
+    (which acts on each input channel separately).
+    The `depth_multiplier` argument controls how many
+    output channels are generated per input channel in the depthwise step.
 
-  Arguments:
-    kernel_size: An integer or tuple/list of 2 integers, specifying the
-        height and width of the 2D convolution window.
-        Can be a single integer to specify the same value for
-        all spatial dimensions.
-    strides: An integer or tuple/list of 2 integers,
-        specifying the strides of the convolution along the height and width.
-        Can be a single integer to specify the same value for
-        all spatial dimensions.
-        Specifying any stride value != 1 is incompatible with specifying
-        any `dilation_rate` value != 1.
-    padding: one of `'valid'` or `'same'` (case-insensitive).
-    depth_multiplier: The number of depthwise convolution output channels
-        for each input channel.
-        The total number of depthwise convolution output
-        channels will be equal to `filters_in * depth_multiplier`.
-    data_format: A string,
-        one of `channels_last` (default) or `channels_first`.
-        The ordering of the dimensions in the inputs.
-        `channels_last` corresponds to inputs with shape
-        `(batch, height, width, channels)` while `channels_first`
-        corresponds to inputs with shape
-        `(batch, channels, height, width)`.
-        It defaults to the `image_data_format` value found in your
-        Keras config file at `~/.keras/keras.json`.
-        If you never set it, then it will be 'channels_last'.
-    activation: Activation function to use.
-        If you don't specify anything, no activation is applied
-        (ie. 'linear' activation: `a(x) = x`).
-    use_bias: Boolean, whether the layer uses a bias vector.
-    depthwise_initializer: Initializer for the depthwise kernel matrix.
-    bias_initializer: Initializer for the bias vector.
-    depthwise_regularizer: Regularizer function applied to
-        the depthwise kernel matrix.
-    bias_regularizer: Regularizer function applied to the bias vector.
-    activity_regularizer: Regularizer function applied to
-        the output of the layer (its 'activation').
-    depthwise_constraint: Constraint function applied to
-        the depthwise kernel matrix.
-    bias_constraint: Constraint function applied to the bias vector.
+    Arguments:
+      kernel_size: An integer or tuple/list of 2 integers, specifying the
+          height and width of the 2D convolution window.
+          Can be a single integer to specify the same value for
+          all spatial dimensions.
+      strides: An integer or tuple/list of 2 integers,
+          specifying the strides of the convolution along the height and width.
+          Can be a single integer to specify the same value for
+          all spatial dimensions.
+          Specifying any stride value != 1 is incompatible with specifying
+          any `dilation_rate` value != 1.
+      padding: one of `'valid'` or `'same'` (case-insensitive).
+      depth_multiplier: The number of depthwise convolution output channels
+          for each input channel.
+          The total number of depthwise convolution output
+          channels will be equal to `filters_in * depth_multiplier`.
+      data_format: A string,
+          one of `channels_last` (default) or `channels_first`.
+          The ordering of the dimensions in the inputs.
+          `channels_last` corresponds to inputs with shape
+          `(batch, height, width, channels)` while `channels_first`
+          corresponds to inputs with shape
+          `(batch, channels, height, width)`.
+          It defaults to the `image_data_format` value found in your
+          Keras config file at `~/.keras/keras.json`.
+          If you never set it, then it will be 'channels_last'.
+      activation: Activation function to use.
+          If you don't specify anything, no activation is applied
+          (ie. 'linear' activation: `a(x) = x`).
+      use_bias: Boolean, whether the layer uses a bias vector.
+      depthwise_initializer: Initializer for the depthwise kernel matrix.
+      bias_initializer: Initializer for the bias vector.
+      depthwise_regularizer: Regularizer function applied to
+          the depthwise kernel matrix.
+      bias_regularizer: Regularizer function applied to the bias vector.
+      activity_regularizer: Regularizer function applied to
+          the output of the layer (its 'activation').
+      depthwise_constraint: Constraint function applied to
+          the depthwise kernel matrix.
+      bias_constraint: Constraint function applied to the bias vector.
 
-  Input shape:
-    4D tensor with shape:
-    `[batch, channels, rows, cols]` if data_format='channels_first'
-    or 4D tensor with shape:
-    `[batch, rows, cols, channels]` if data_format='channels_last'.
+    Input shape:
+      4D tensor with shape:
+      `[batch, channels, rows, cols]` if data_format='channels_first'
+      or 4D tensor with shape:
+      `[batch, rows, cols, channels]` if data_format='channels_last'.
 
-  Output shape:
-    4D tensor with shape:
-    `[batch, filters, new_rows, new_cols]` if data_format='channels_first'
-    or 4D tensor with shape:
-    `[batch, new_rows, new_cols, filters]` if data_format='channels_last'.
-    `rows` and `cols` values might have changed due to padding.
-  """
+    Output shape:
+      4D tensor with shape:
+      `[batch, filters, new_rows, new_cols]` if data_format='channels_first'
+      or 4D tensor with shape:
+      `[batch, new_rows, new_cols, filters]` if data_format='channels_last'.
+      `rows` and `cols` values might have changed due to padding.
+    """
 
     def __init__(
         self,
@@ -433,19 +437,29 @@ class DepthwiseConv2D(Conv2D):
 
         # Not implemented arguments
         default_args_check(
-            depthwise_regularizer, "depthwise_regularizer", "DepthwiseConv2D",
+            depthwise_regularizer,
+            "depthwise_regularizer",
+            "DepthwiseConv2D",
         )
         default_args_check(
-            bias_regularizer, "bias_regularizer", "DepthwiseConv2D",
+            bias_regularizer,
+            "bias_regularizer",
+            "DepthwiseConv2D",
         )
         default_args_check(
-            activity_regularizer, "activity_regularizer", "DepthwiseConv2D",
+            activity_regularizer,
+            "activity_regularizer",
+            "DepthwiseConv2D",
         )
         default_args_check(
-            depthwise_constraint, "depthwise_constraint", "DepthwiseConv2D",
+            depthwise_constraint,
+            "depthwise_constraint",
+            "DepthwiseConv2D",
         )
         default_args_check(
-            bias_constraint, "bias_constraint", "DepthwiseConv2D",
+            bias_constraint,
+            "bias_constraint",
+            "DepthwiseConv2D",
         )
 
     def build(self, input_shape):
@@ -479,9 +493,9 @@ class DepthwiseConv2D(Conv2D):
     def rearrange_kernel(self, kernel):
         """Rearrange kernel to match normal convolution kernels
 
-    Arguments:
-      kernel: kernel to be rearranged
-    """
+        Arguments:
+          kernel: kernel to be rearranged
+        """
         mask = self.get_mask(self.input_dim)
 
         if isinstance(kernel, tf.Tensor):
@@ -500,8 +514,7 @@ class DepthwiseConv2D(Conv2D):
                 # rearrange kernel
                 kernel = tf.reshape(
                     kernel,
-                    shape=self.kernel_size
-                    + (1, self.input_dim * self.depth_multiplier),
+                    self.kernel_size + (1, self.input_dim * self.depth_multiplier),
                 )
 
             kernel = tf.multiply(kernel, mask)
@@ -511,8 +524,7 @@ class DepthwiseConv2D(Conv2D):
                 # rearrange kernel
                 kernel = np.reshape(
                     kernel,
-                    newshape=self.kernel_size
-                    + (1, self.input_dim * self.depth_multiplier),
+                    self.kernel_size + (1, self.input_dim * self.depth_multiplier),
                 )
 
             kernel = np.multiply(kernel, mask)
@@ -523,8 +535,7 @@ class DepthwiseConv2D(Conv2D):
                 # rearrange kernel
                 kernel = tfe.reshape(
                     kernel,
-                    shape=self.kernel_size
-                    + (1, self.input_dim * self.depth_multiplier),
+                    self.kernel_size + (1, self.input_dim * self.depth_multiplier),
                 )
 
             kernel = tfe.mul(kernel, mask)
@@ -538,10 +549,12 @@ class DepthwiseConv2D(Conv2D):
         if self.data_format != "channels_first":
             inputs = tfe.transpose(inputs, perm=[0, 3, 1, 2])
 
-        outputs = tfe.conv2d(inputs, self.kernel, self.strides[0], self.padding)
+        outputs = tfe.conv2d(
+            inputs, self.kernel.read_value(), self.strides, self.padding
+        )
 
         if self.use_bias:
-            outputs = outputs + self.bias
+            outputs = outputs + self.bias.read_value()
 
         if self.data_format != "channels_first":
             outputs = tfe.transpose(outputs, perm=[0, 2, 3, 1])
@@ -558,9 +571,9 @@ class DepthwiseConv2D(Conv2D):
         h_filter, w_filter, _, n_filters = self.kernel_shape
 
         if self.data_format == "channels_first":
-            n_x, _, h_x, w_x = input_shape.as_list()
+            n_x, _, h_x, w_x = input_shape
         else:
-            n_x, h_x, w_x, _ = input_shape.as_list()
+            n_x, h_x, w_x, _ = input_shape
 
         if self.padding == "SAME":
             h_out = int(np.ceil(float(h_x) / float(self.strides[0])))
@@ -585,50 +598,36 @@ class DepthwiseConv2D(Conv2D):
                 mask[:, :, i, d + (i * self.depth_multiplier)] = 1.0
         return mask
 
-    def set_weights(self, weights, sess=None):
+    def set_weights(self, weights):
         """
-    Sets the weights of the layer.
+        Sets the weights of the layer.
 
-    Arguments:
-      weights: A list of Numpy arrays with shapes and types
-          matching the output of layer.get_weights() or a list
-          of private variables
-      sess: tfe session"""
+        Arguments:
+          weights: A list of Numpy arrays with shapes and types
+              matching the output of layer.get_weights() or a list
+              of private variables
+          sess: tfe session"""
 
         weights_types = (np.ndarray, TFEPrivateTensor)
         assert isinstance(weights[0], weights_types), type(weights[0])
 
         # Assign new keras weights to existing weights defined by
         # default when tfe layer was instantiated
-        if not sess:
-            sess = KE.get_session()
-
         if isinstance(weights[0], np.ndarray):
-            for i, w in enumerate(self.weights):
-                shape = w.shape.as_list()
-                tfe_weights_pl = tfe.define_private_placeholder(shape)
+            for index, weight in enumerate(weights):
+                weights[index] = tfe.define_private_variable(weight)
+        if isinstance(weights[0], TFEPrivateVariable):
+            for index, weight in enumerate(weights):
+                weights[index] = weights[index].read_value()
 
-                new_weight = weights[i]
-                if i == 0:
-                    # kernel
-                    new_weight = self.rearrange_kernel(new_weight)
-                else:
-                    # bias
-                    new_weight = new_weight.reshape(shape)
+        for i, w in enumerate(self.weights):
+            shape = w.shape.as_list()
+            new_weight = weights[i]
+            if i == 0:
+                # kernel
+                new_weight = self.rearrange_kernel(new_weight)
+            else:
+                # bias
+                new_weight = new_weight.reshape(shape)
 
-                fd = tfe_weights_pl.feed(new_weight)
-                sess.run(tfe.assign(w, tfe_weights_pl), feed_dict=fd)
-
-        elif isinstance(weights[0], TFEPrivateTensor):
-            for i, w in enumerate(self.weights):
-                shape = w.shape.as_list()
-
-                new_weight = weights[i]
-                if i == 0:
-                    # kernel
-                    new_weight = self.rearrange_kernel(new_weight)
-                else:
-                    # bias
-                    new_weight = new_weight.reshape(shape)
-
-                sess.run(tfe.assign(w, new_weight))
+            tfe.assign(w, new_weight)
