@@ -6,7 +6,6 @@ import pytest
 import tensorflow as tf
 
 import tf_encrypted as tfe
-from tf_encrypted.keras import backend as KE
 from tf_encrypted.keras.testing_utils import agreement_test
 from tf_encrypted.keras.testing_utils import layer_test
 
@@ -15,9 +14,6 @@ np.random.seed(42)
 
 @pytest.mark.layers
 class TestConv2d(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_conv2d_bias(self):
         self._core_conv2d(kernel_size=2, use_bias=True)
 
@@ -32,7 +28,7 @@ class TestConv2d(unittest.TestCase):
 
     def _core_conv2d(self, **layer_kwargs):
         filters_in = 3
-        input_shape = [2, 6, 6, filters_in]  # channels last
+        batch_input_shape = [2, 6, 6, filters_in]  # channels last
         filters = 5
 
         if isinstance(layer_kwargs["kernel_size"], int):
@@ -50,17 +46,16 @@ class TestConv2d(unittest.TestCase):
         }
 
         kwargs = {**base_kwargs, **layer_kwargs}
-        agreement_test(tfe.keras.layers.Conv2D, kwargs=kwargs, input_shape=input_shape)
+        agreement_test(
+            tfe.keras.layers.Conv2D, kwargs=kwargs, batch_input_shape=batch_input_shape
+        )
         layer_test(
-            tfe.keras.layers.Conv2D, kwargs=kwargs, batch_input_shape=input_shape
+            tfe.keras.layers.Conv2D, kwargs=kwargs, batch_input_shape=batch_input_shape
         )
 
 
 @pytest.mark.layers
 class TestDepthwiseConv2d(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_depthwise_conv2d_bias(self):
         self._core_depthwise_conv2d(kernel_size=2, use_bias=True)
 
@@ -80,18 +75,18 @@ class TestDepthwiseConv2d(unittest.TestCase):
         input_shape = (1, 10, 10, 3)
         input_data = np.random.normal(size=input_shape)
 
-        with tf.Session():
-            model = tf.keras.models.Sequential()
+        model = tf.keras.models.Sequential()
 
-            model.add(
-                tf.keras.layers.DepthwiseConv2D(
-                    kernel_size=(2, 2), batch_input_shape=input_shape,
-                )
+        model.add(
+            tf.keras.layers.DepthwiseConv2D(
+                kernel_size=(2, 2),
+                batch_input_shape=input_shape,
             )
+        )
 
-            expected = model.predict(input_data)
-            k_weights = model.get_weights()
-            k_config = model.get_config()
+        expected = model.predict(input_data)
+        k_weights = model.get_weights()
+        k_config = model.get_config()
 
         with tf.name_scope("TFE"):
             x = tfe.define_private_input(
@@ -101,17 +96,13 @@ class TestDepthwiseConv2d(unittest.TestCase):
             tfe_model = tfe.keras.models.model_from_config(k_config)
             tfe_model.set_weights(k_weights)
             y = tfe_model(x)
+            actual = y.reveal().to_native()
 
-        with KE.get_session() as sess:
-            actual = sess.run(y.reveal())
-
-            np.testing.assert_allclose(actual, expected, rtol=1e-2, atol=1e-2)
-
-        KE.clear_session()
+        np.testing.assert_allclose(actual, expected, rtol=1e-2, atol=1e-2)
 
     def _core_depthwise_conv2d(self, **layer_kwargs):
         filters_in = 3
-        input_shape = [2, 6, 6, filters_in]  # channels last
+        batch_input_shape = [2, 6, 6, filters_in]  # channels last
 
         if isinstance(layer_kwargs["kernel_size"], int):
             kernel_size_in = (layer_kwargs["kernel_size"],) * 2
@@ -120,7 +111,7 @@ class TestDepthwiseConv2d(unittest.TestCase):
 
         filters_out = layer_kwargs.get("depth_multiplier", 1)
 
-        kernel = np.random.normal(kernel_size_in + (filters_in, filters_out))
+        kernel = np.random.normal(size=kernel_size_in + (filters_in, filters_out))
 
         initializer = tf.keras.initializers.Constant(kernel)
 
@@ -133,13 +124,13 @@ class TestDepthwiseConv2d(unittest.TestCase):
         agreement_test(
             tfe.keras.layers.DepthwiseConv2D,
             kwargs=kwargs,
-            input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
             atol=1e-2,
         )
         layer_test(
             tfe.keras.layers.DepthwiseConv2D,
             kwargs=kwargs,
-            batch_input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
 
 
