@@ -11,7 +11,7 @@ class UnknownLayerArgError(ValueError):
       arg_name: TF Encrypted Keras layer argument name (string)
       layer_sign: TensorFlow Keras layer signature (dict)
       tf_layer_name: TensorFlow Keras layer name (string)
-  """
+    """
 
     def __init__(self, arg_name, layer_sign, layer_name):
         super(UnknownLayerArgError, self).__init__()
@@ -38,7 +38,7 @@ class LayerArgNotImplementedError(NotImplementedError):
       arg: TFE layer argument
       arg_name: TFE layer argument name (string)
       tf_layer_name: Tensorflow keras layer name (string)
-  """
+    """
 
     def __init__(self, arg_name, tf_layer_name, tf_default_arg):
         super(LayerArgNotImplementedError, self).__init__()
@@ -57,7 +57,7 @@ class LayerArgNotImplementedError(NotImplementedError):
 
 
 def default_args_check(arg, arg_name, tf_layer_name):
-    """Check if the layer is using the dfault argument
+    """Check if the layer is using the default argument
 
     Args:
       arg: TFE layer argument
@@ -66,7 +66,7 @@ def default_args_check(arg, arg_name, tf_layer_name):
 
     Raises:
       NotImplementedError: if this argument is not implemented for this `layer`.
-  """
+    """
     tf_layer_cls = getattr(tf.keras.layers, tf_layer_name)
     layer_sign = inspect.signature(tf_layer_cls.__init__).parameters
     if arg_name not in layer_sign:
