@@ -14,9 +14,6 @@ np.random.seed(42)
 
 @pytest.mark.layers
 class TestBatchNormalization(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_batchnorm_2d(self):
         self._core_batchnorm([1, 4], axis=1)
 
@@ -38,7 +35,8 @@ class TestBatchNormalization(unittest.TestCase):
         initializer = tf.keras.initializers.Constant(const)
 
         self._core_batchnorm(
-            [1] + input_shape, moving_mean_initializer=initializer,
+            [1] + input_shape,
+            moving_mean_initializer=initializer,
         )
 
     def test_batchnorm_non_default_variance_init(self):
@@ -47,21 +45,24 @@ class TestBatchNormalization(unittest.TestCase):
         initializer = tf.keras.initializers.Constant(const)
 
         self._core_batchnorm(
-            [1] + input_shape, moving_variance_initializer=initializer,
+            [1] + input_shape,
+            moving_variance_initializer=initializer,
         )
 
-    def _core_batchnorm(self, input_shape, **layer_kwargs):
+    def _core_batchnorm(self, batch_input_shape, **layer_kwargs):
         base_kwargs = {"fused": False}
 
         kwargs = {**base_kwargs, **layer_kwargs}
 
         agreement_test(
-            tfe.keras.layers.BatchNormalization, kwargs=kwargs, input_shape=input_shape,
+            tfe.keras.layers.BatchNormalization,
+            kwargs=kwargs,
+            batch_input_shape=batch_input_shape,
         )
         layer_test(
             tfe.keras.layers.BatchNormalization,
             kwargs=kwargs,
-            batch_input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
 
 
