@@ -10,25 +10,25 @@ from tf_encrypted.keras.engine import Layer
 
 class Pooling2D(Layer):
     """Pooling layer for arbitrary pooling functions, for 2D inputs (e.g. images).
-  This class only exists for code reuse. It will never be an exposed API.
-  Arguments:
-    _pool_function: The pooling function to apply, e.g. `prot.max_pool2d`.
-    pool_size: An integer or tuple/list of 2 integers: (pool_height, pool_width)
-      specifying the size of the pooling window.
-      Can be a single integer to specify the same value for
-      all spatial dimensions.
-    strides: An integer or tuple/list of 2 integers,
-      specifying the strides of the pooling operation.
-      Can be a single integer to specify the same value for
-      all spatial dimensions.
-    padding: A string. The padding method, either 'valid' or 'same'.
-      Case-insensitive.
-    data_format: A string, one of `channels_last` (default) or `channels_first`.
-      The ordering of the dimensions in the inputs.
-      `channels_last` corresponds to inputs with shape
-      `(batch, height, width, channels)` while `channels_first` corresponds to
-      inputs with shape `(batch, channels, height, width)`.
-  """
+    This class only exists for code reuse. It will never be an exposed API.
+    Arguments:
+      _pool_function: The pooling function to apply, e.g. `prot.max_pool2d`.
+      pool_size: An integer or tuple/list of 2 integers: (pool_height, pool_width)
+        specifying the size of the pooling window.
+        Can be a single integer to specify the same value for
+        all spatial dimensions.
+      strides: An integer or tuple/list of 2 integers,
+        specifying the strides of the pooling operation.
+        Can be a single integer to specify the same value for
+        all spatial dimensions.
+      padding: A string. The padding method, either 'valid' or 'same'.
+        Case-insensitive.
+      data_format: A string, one of `channels_last` (default) or `channels_first`.
+        The ordering of the dimensions in the inputs.
+        `channels_last` corresponds to inputs with shape
+        `(batch, height, width, channels)` while `channels_first` corresponds to
+        inputs with shape `(batch, channels, height, width)`.
+    """
 
     def __init__(
         self,
@@ -89,38 +89,38 @@ class Pooling2D(Layer):
 
 class MaxPooling2D(Pooling2D):
     """Max pooling operation for spatial data.
-  Arguments:
-      pool_size: integer or tuple of 2 integers,
-          factors by which to downscale (vertical, horizontal).
-          (2, 2) will halve the input in both spatial dimension.
-          If only one integer is specified, the same window length
-          will be used for both dimensions.
-      strides: Integer, tuple of 2 integers, or None.
-          Strides values.
-          If None, it will default to `pool_size`.
-      padding: One of `"valid"` or `"same"` (case-insensitive).
-      data_format: A string,
-          one of `channels_last` (default) or `channels_first`.
-          The ordering of the dimensions in the inputs.
-          `channels_last` corresponds to inputs with shape
-          `(batch, height, width, channels)` while `channels_first`
-          corresponds to inputs with shape
-          `(batch, channels, height, width)`.
-  Input shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, rows, cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, rows, cols)`
-  Output shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, pooled_rows, pooled_cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, pooled_rows, pooled_cols)`
-  """
+    Arguments:
+        pool_size: integer or tuple of 2 integers,
+            factors by which to downscale (vertical, horizontal).
+            (2, 2) will halve the input in both spatial dimension.
+            If only one integer is specified, the same window length
+            will be used for both dimensions.
+        strides: Integer, tuple of 2 integers, or None.
+            Strides values.
+            If None, it will default to `pool_size`.
+        padding: One of `"valid"` or `"same"` (case-insensitive).
+        data_format: A string,
+            one of `channels_last` (default) or `channels_first`.
+            The ordering of the dimensions in the inputs.
+            `channels_last` corresponds to inputs with shape
+            `(batch, height, width, channels)` while `channels_first`
+            corresponds to inputs with shape
+            `(batch, channels, height, width)`.
+    Input shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, rows, cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, rows, cols)`
+    Output shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, pooled_rows, pooled_cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, pooled_rows, pooled_cols)`
+    """
 
     def __init__(
         self,
@@ -177,7 +177,7 @@ class MaxPooling2D(Pooling2D):
         d_x = tfe.patches2im(
             _d_y,
             self.pool_size,
-            stride=self.strides[0],
+            strides=self.strides,
             padding=self.padding,
             img_size=(h_x, w_x),
             consolidation="SUM",
@@ -193,38 +193,38 @@ class MaxPooling2D(Pooling2D):
 
 class AveragePooling2D(Pooling2D):
     """Average pooling operation for spatial data.
-  Arguments:
-      pool_size: integer or tuple of 2 integers,
-          factors by which to downscale (vertical, horizontal).
-          (2, 2) will halve the input in both spatial dimension.
-          If only one integer is specified, the same window length
-          will be used for both dimensions.
-      strides: Integer, tuple of 2 integers, or None.
-          Strides values.
-          If None, it will default to `pool_size`.
-      padding: One of `"valid"` or `"same"` (case-insensitive).
-      data_format: A string,
-          one of `channels_last` (default) or `channels_first`.
-          The ordering of the dimensions in the inputs.
-          `channels_last` corresponds to inputs with shape
-          `(batch, height, width, channels)` while `channels_first`
-          corresponds to inputs with shape
-          `(batch, channels, height, width)`.
-  Input shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, rows, cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, rows, cols)`
-  Output shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, pooled_rows, pooled_cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, pooled_rows, pooled_cols)`
-  """
+    Arguments:
+        pool_size: integer or tuple of 2 integers,
+            factors by which to downscale (vertical, horizontal).
+            (2, 2) will halve the input in both spatial dimension.
+            If only one integer is specified, the same window length
+            will be used for both dimensions.
+        strides: Integer, tuple of 2 integers, or None.
+            Strides values.
+            If None, it will default to `pool_size`.
+        padding: One of `"valid"` or `"same"` (case-insensitive).
+        data_format: A string,
+            one of `channels_last` (default) or `channels_first`.
+            The ordering of the dimensions in the inputs.
+            `channels_last` corresponds to inputs with shape
+            `(batch, height, width, channels)` while `channels_first`
+            corresponds to inputs with shape
+            `(batch, channels, height, width)`.
+    Input shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, rows, cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, rows, cols)`
+    Output shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, pooled_rows, pooled_cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, pooled_rows, pooled_cols)`
+    """
 
     def __init__(
         self,
@@ -278,7 +278,7 @@ class AveragePooling2D(Pooling2D):
         d_x = tfe.patches2im(
             _d_y,
             self.pool_size,
-            stride=self.strides[0],
+            strides=self.strides,
             padding=self.padding,
             img_size=(h_x, w_x),
             consolidation="SUM",
@@ -293,18 +293,24 @@ class AveragePooling2D(Pooling2D):
 
 
 class GlobalPooling2D(Layer):
-    """Abstract class for different global pooling 2D layers.
-  """
+    """Abstract class for different global pooling 2D layers."""
 
-    def __init__(self, data_format=None, **kwargs):
+    def __init__(self, data_format=None, keepdims=False, **kwargs):
         super(GlobalPooling2D, self).__init__(**kwargs)
         self.data_format = conv_utils.normalize_data_format(data_format)
+        self.keepdims = keepdims
 
     def compute_output_shape(self, input_shape):
         if self.data_format == "channels_last":
-            output_shape = [input_shape[0], input_shape[3]]
+            if self.keepdims:
+                output_shape = [input_shape[0], 1, 1, input_shape[3]]
+            else:
+                output_shape = [input_shape[0], input_shape[3]]
         else:
-            output_shape = [input_shape[0], input_shape[1]]
+            if self.keepdims:
+                output_shape = [input_shape[0], input_shape[1], 1, 1]
+            else:
+                output_shape = [input_shape[0], input_shape[1]]
 
         return output_shape
 
@@ -316,30 +322,30 @@ class GlobalPooling2D(Layer):
 class GlobalAveragePooling2D(GlobalPooling2D):
     """Global average pooling operation for spatial data.
 
-  Arguments:
-      data_format: A string,
-          one of `channels_last` (default) or `channels_first`.
-          The ordering of the dimensions in the inputs.
-          `channels_last` corresponds to inputs with shape
-          `(batch, height, width, channels)` while `channels_first`
-          corresponds to inputs with shape
-          `(batch, channels, height, width)`.
-          It defaults to the `image_data_format` value found in your
-          Keras config file at `~/.keras/keras.json`.
-          If you never set it, then it will be "channels_last".
+    Arguments:
+        data_format: A string,
+            one of `channels_last` (default) or `channels_first`.
+            The ordering of the dimensions in the inputs.
+            `channels_last` corresponds to inputs with shape
+            `(batch, height, width, channels)` while `channels_first`
+            corresponds to inputs with shape
+            `(batch, channels, height, width)`.
+            It defaults to the `image_data_format` value found in your
+            Keras config file at `~/.keras/keras.json`.
+            If you never set it, then it will be "channels_last".
 
-  Input shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, rows, cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, rows, cols)`
+    Input shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, rows, cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, rows, cols)`
 
-  Output shape:
-      2D tensor with shape:
-      `(batch_size, channels)`
-  """
+    Output shape:
+        2D tensor with shape:
+        `(batch_size, channels)`
+    """
 
     def build(self, input_shape):
         if self.data_format == "channels_last":
@@ -356,7 +362,11 @@ class GlobalAveragePooling2D(GlobalPooling2D):
         else:
             x_reduced = inputs.reduce_sum(axis=3).reduce_sum(axis=2)
 
-        return x_reduced * self.scalar
+        output_shape = self.compute_output_shape(inputs.shape)
+        res = x_reduced * self.scalar
+        res = res.reshape(output_shape)
+
+        return res
 
     def backward(self, d_y):
         x = self._layer_input
@@ -366,7 +376,8 @@ class GlobalAveragePooling2D(GlobalPooling2D):
             h_x, w_x = (x.shape[2], x.shape[3])
 
         _d_y = d_y * self.scalar
-        _d_y = tfe.expand_dims(tfe.expand_dims(_d_y, axis=2), axis=3)
+        if not self.keepdims:
+            _d_y = tfe.expand_dims(tfe.expand_dims(_d_y, axis=2), axis=3)
 
         d_x = tfe.tile(_d_y, multiples=[1, 1, h_x, w_x])
 
@@ -379,30 +390,30 @@ class GlobalAveragePooling2D(GlobalPooling2D):
 class GlobalMaxPooling2D(GlobalPooling2D):
     """Global max pooling operation for spatial data.
 
-  Arguments:
-      data_format: A string,
-          one of `channels_last` (default) or `channels_first`.
-          The ordering of the dimensions in the inputs.
-          `channels_last` corresponds to inputs with shape
-          `(batch, height, width, channels)` while `channels_first`
-          corresponds to inputs with shape
-          `(batch, channels, height, width)`.
-          It defaults to the `image_data_format` value found in your
-          Keras config file at `~/.keras/keras.json`.
-          If you never set it, then it will be "channels_last".
+    Arguments:
+        data_format: A string,
+            one of `channels_last` (default) or `channels_first`.
+            The ordering of the dimensions in the inputs.
+            `channels_last` corresponds to inputs with shape
+            `(batch, height, width, channels)` while `channels_first`
+            corresponds to inputs with shape
+            `(batch, channels, height, width)`.
+            It defaults to the `image_data_format` value found in your
+            Keras config file at `~/.keras/keras.json`.
+            If you never set it, then it will be "channels_last".
 
-  Input shape:
-      - If `data_format='channels_last'`:
-          4D tensor with shape:
-          `(batch_size, rows, cols, channels)`
-      - If `data_format='channels_first'`:
-          4D tensor with shape:
-          `(batch_size, channels, rows, cols)`
+    Input shape:
+        - If `data_format='channels_last'`:
+            4D tensor with shape:
+            `(batch_size, rows, cols, channels)`
+        - If `data_format='channels_first'`:
+            4D tensor with shape:
+            `(batch_size, channels, rows, cols)`
 
-  Output shape:
-      2D tensor with shape:
-      `(batch_size, channels)`
-  """
+    Output shape:
+        2D tensor with shape:
+        `(batch_size, channels)`
+    """
 
     def call(self, inputs):
         self._layer_input = inputs
