@@ -3,7 +3,6 @@ import unittest
 
 import numpy as np
 import pytest
-import tensorflow as tf
 
 import tf_encrypted as tfe
 from tf_encrypted.keras.testing_utils import agreement_test
@@ -14,9 +13,6 @@ np.random.seed(42)
 
 @pytest.mark.layers
 class TestPooling2d(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_maxpooling2d_valid(self):
         self._core_maxpooling2d(strides=2, padding="valid")
 
@@ -37,7 +33,7 @@ class TestPooling2d(unittest.TestCase):
 
     def _core_maxpooling2d(self, **layer_kwargs):
         channel_in = 2
-        input_shape = [2, 8, 8, channel_in]  # channels last
+        batch_input_shape = [2, 8, 8, channel_in]  # channels last
         pool_size_in = 2
 
         base_kwargs = {
@@ -46,15 +42,19 @@ class TestPooling2d(unittest.TestCase):
 
         kwargs = {**base_kwargs, **layer_kwargs}
         agreement_test(
-            tfe.keras.layers.MaxPooling2D, kwargs=kwargs, input_shape=input_shape,
+            tfe.keras.layers.MaxPooling2D,
+            kwargs=kwargs,
+            batch_input_shape=batch_input_shape,
         )
         layer_test(
-            tfe.keras.layers.MaxPooling2D, kwargs=kwargs, batch_input_shape=input_shape,
+            tfe.keras.layers.MaxPooling2D,
+            kwargs=kwargs,
+            batch_input_shape=batch_input_shape,
         )
 
     def _core_avgpooling2d(self, **layer_kwargs):
         channel_in = 2
-        input_shape = [2, 8, 8, channel_in]  # channels last
+        batch_input_shape = [2, 8, 8, channel_in]  # channels last
         pool_size_in = 2
 
         base_kwargs = {
@@ -63,20 +63,19 @@ class TestPooling2d(unittest.TestCase):
 
         kwargs = {**base_kwargs, **layer_kwargs}
         agreement_test(
-            tfe.keras.layers.AveragePooling2D, kwargs=kwargs, input_shape=input_shape,
+            tfe.keras.layers.AveragePooling2D,
+            kwargs=kwargs,
+            batch_input_shape=batch_input_shape,
         )
         layer_test(
             tfe.keras.layers.AveragePooling2D,
             kwargs=kwargs,
-            batch_input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
 
 
 @pytest.mark.layers
 class TestGlobalPooling2d(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_global_maxpooling2d(self):
         self._core_global_maxpooling2d()
 
@@ -85,23 +84,25 @@ class TestGlobalPooling2d(unittest.TestCase):
 
     def _core_global_maxpooling2d(self, **layer_kwargs):
         channel_in = 2
-        input_shape = [2, 4, 4, channel_in]  # channels last
+        batch_input_shape = [2, 4, 4, channel_in]  # channels last
 
         base_kwargs = {}
 
         kwargs = {**base_kwargs, **layer_kwargs}
         agreement_test(
-            tfe.keras.layers.GlobalMaxPooling2D, kwargs=kwargs, input_shape=input_shape,
+            tfe.keras.layers.GlobalMaxPooling2D,
+            kwargs=kwargs,
+            batch_input_shape=batch_input_shape,
         )
         layer_test(
             tfe.keras.layers.GlobalMaxPooling2D,
             kwargs=kwargs,
-            batch_input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
 
     def _core_global_avgpooling2d(self, **layer_kwargs):
         channel_in = 2
-        input_shape = [2, 4, 4, channel_in]  # channels last
+        batch_input_shape = [2, 4, 4, channel_in]  # channels last
 
         base_kwargs = {}
 
@@ -109,12 +110,12 @@ class TestGlobalPooling2d(unittest.TestCase):
         agreement_test(
             tfe.keras.layers.GlobalAveragePooling2D,
             kwargs=kwargs,
-            input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
         layer_test(
             tfe.keras.layers.GlobalAveragePooling2D,
             kwargs=kwargs,
-            batch_input_shape=input_shape,
+            batch_input_shape=batch_input_shape,
         )
 
 
