@@ -71,7 +71,12 @@ def supports_seeded_randomness():
 
 
 def seeded_random_uniform(
-    shape, minval=0, maxval=None, dtype=tf.int32, seed=None, name=None,
+    shape,
+    minval=0,
+    maxval=None,
+    dtype=tf.int32,
+    seed=None,
+    name=None,
 ):
     """
     Returns cryptographically strong random numbers with a seed
@@ -103,12 +108,20 @@ def seeded_random_uniform(
     maxval = ops.convert_to_tensor(maxval, dtype=dtype, name="max")
 
     return secure_random_module.secure_seeded_random_uniform(
-        shape, seed, minval, maxval, name=name,
+        shape,
+        seed,
+        minval,
+        maxval,
+        name=name,
     )
 
 
 def random_uniform(
-    shape, minval=0, maxval=None, dtype=tf.int32, name=None,
+    shape,
+    minval=0,
+    maxval=None,
+    dtype=tf.int32,
+    name=None,
 ):
     """
     Returns cryptographically strong random numbers.
@@ -137,7 +150,12 @@ def random_uniform(
     minval = ops.convert_to_tensor(minval, dtype=dtype, name="min")
     maxval = ops.convert_to_tensor(maxval, dtype=dtype, name="max")
 
-    return secure_random_module.secure_random_uniform(shape, minval, maxval, name=name,)
+    return secure_random_module.secure_random_uniform(
+        shape,
+        minval,
+        maxval,
+        name=name,
+    )
 
 
 def secure_seed(name=None):
