@@ -1,8 +1,7 @@
 """The TF Encrypted Player abstraction."""
 from __future__ import absolute_import
 
+from .data_owner import DataOwner
 from .player import Player
 
-__all__ = [
-    "Player",
-]
+__all__ = ["Player", "DataOwner"]
