@@ -4,23 +4,27 @@ from __future__ import absolute_import
 from .pond import AdditiveFIFOQueue
 from .pond import Pond
 from .pond import PondMaskedTensor
+from .pond import PondMaskedVariable
 from .pond import PondPrivateTensor
 from .pond import PondPrivateVariable
 from .pond import PondPublicTensor
+from .pond import PondPublicVariable
 from .pond import PondTensor
 from .pond import TFEInputter
 from .pond import _type
 from .triple_sources import OnlineTripleSource
-from .triple_sources import QueuedOnlineTripleSource
+
+# from .triple_sources import QueuedOnlineTripleSource
 
 __all__ = [
     "Pond",
-    "PondPublicTensor",
     "PondTensor",
     "PondPublicTensor",
     "PondPrivateTensor",
-    "PondPrivateVariable",
     "PondMaskedTensor",
+    "PondPublicVariable",
+    "PondPrivateVariable",
+    "PondMaskedVariable",
     "TFEInputter",
     "_type",
     "OnlineTripleSource",
