@@ -4,35 +4,14 @@ import unittest
 import numpy as np
 import tensorflow as tf
 
-import tf_encrypted as tfe
-from tf_encrypted.tensor import fixed64
 from tf_encrypted.tensor import int64factory
 
 
 class TestInt64Tensor(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
-    def test_pond(self) -> None:
-
-        with tfe.protocol.Pond(
-            tensor_factory=int64factory, fixedpoint_config=fixed64,
-        ) as prot:
-
-            x = prot.define_private_variable(np.array([2, 2]), apply_scaling=False)
-            y = prot.define_public_variable(np.array([2, 2]), apply_scaling=False)
-
-            z = x * y
-
-            with tfe.Session() as sess:
-                sess.run(tf.global_variables_initializer())
-                out = sess.run(z.reveal())
-                np.testing.assert_array_almost_equal(out, [4, 4], decimal=3)
-
     def test_binarize(self) -> None:
         x = int64factory.tensor(
             tf.constant(
-                [2 ** 62 + 3, 2 ** 63 - 1, 2 ** 63 - 2, -3],
+                [2**62 + 3, 2**63 - 1, 2**63 - 2, -3],
                 shape=[2, 2],
                 dtype=tf.int64,
             )
@@ -61,23 +40,23 @@ class TestInt64Tensor(unittest.TestCase):
         ]).reshape([2, 2, 64])
         # fmt: on
 
-        with tf.Session() as sess:
-            actual = sess.run(y.to_native())
-
+        actual = y.to_native()
         np.testing.assert_array_equal(actual, expected)
 
     def test_random_binarize(self) -> None:
         x_in = (
-            np.random.uniform(low=2 ** 63 + 1, high=2 ** 63 - 1, size=2000,)
+            np.random.uniform(
+                low=2**63 + 1,
+                high=2**63 - 1,
+                size=2000,
+            )
             .astype(np.int64)
             .tolist()
         )
         x = int64factory.tensor(tf.constant(x_in, dtype=tf.int64))
 
         y = x.bits()
-
-        with tf.Session() as sess:
-            actual = sess.run(y.to_native())
+        actual = y.to_native()
 
         j = 0
         for i in x_in:
@@ -92,9 +71,6 @@ class TestInt64Tensor(unittest.TestCase):
 
 
 class TestConv2D(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def test_forward(self) -> None:
         # input
         batch_size, channels_in, channels_out = 32, 3, 64
@@ -103,34 +79,30 @@ class TestConv2D(unittest.TestCase):
         input_conv = np.random.normal(size=input_shape).astype(np.int64)
 
         # filters
-        h_filter, w_filter, strides = 2, 2, 2
+        h_filter, w_filter = 2, 2
+        strides = [2, 2]
         filter_shape = (h_filter, w_filter, channels_in, channels_out)
         filter_values = np.random.normal(size=filter_shape).astype(np.int64)
 
         x_in = int64factory.tensor(input_conv)
         out = x_in.conv2d(int64factory.tensor(filter_values), strides)
-        with tf.Session() as sess:
-            actual = sess.run(out.to_native())
+        actual = out.to_native()
 
-        # reset graph
-        tf.reset_default_graph()
+        # conv input
+        x = tf.Variable(input_conv, dtype=tf.float32)
+        x_nhwc = tf.transpose(x, (0, 2, 3, 1))
 
-        # convolution tensorflow
-        with tf.Session() as sess:
-            # conv input
-            x = tf.Variable(input_conv, dtype=tf.float32)
-            x_nhwc = tf.transpose(x, (0, 2, 3, 1))
+        # conv filter
+        filters_tf = tf.Variable(filter_values, dtype=tf.float32)
 
-            # convolution Tensorflow
-            filters_tf = tf.Variable(filter_values, dtype=tf.float32)
+        conv_out_tf = tf.nn.conv2d(
+            x_nhwc,
+            filters_tf,
+            strides=[1, strides[0], strides[1], 1],
+            padding="SAME",
+        )
 
-            conv_out_tf = tf.nn.conv2d(
-                x_nhwc, filters_tf, strides=[1, strides, strides, 1], padding="SAME",
-            )
-
-            sess.run(tf.global_variables_initializer())
-            out_tensorflow = sess.run(conv_out_tf).transpose(0, 3, 1, 2)
-
+        out_tensorflow = tf.transpose(conv_out_tf, perm=[0, 3, 1, 2])
         np.testing.assert_array_almost_equal(actual, out_tensorflow, decimal=3)
 
 
