@@ -8,11 +8,8 @@ from tf_encrypted.tensor import native_factory
 
 
 class TestPrimeTensor(unittest.TestCase):
-    def setUp(self):
-        tf.reset_default_graph()
-
     def set_up_indexing(self):
-        m = 2 ** 31
+        m = 2**31
 
         prime_factory = native_factory(tf.int32, m)
 
@@ -45,34 +42,43 @@ class TestPrimeTensor(unittest.TestCase):
             ndim = len(np_fix.shape)
             if ndim == 1:
                 np.testing.assert_equal(
-                    np_fix[2:5], prime_fix[2:5].value,
+                    np_fix[2:5],
+                    prime_fix[2:5].value,
                 )
                 continue
             np.testing.assert_equal(
-                np_fix[:, 0], prime_fix[:, 0].value,
+                np_fix[:, 0],
+                prime_fix[:, 0].value,
             )
             np.testing.assert_equal(
-                np_fix[:, 1], prime_fix[:, 1].value,
+                np_fix[:, 1],
+                prime_fix[:, 1].value,
             )
             np.testing.assert_equal(
-                np_fix[:, -1], prime_fix[:, -1].value,
+                np_fix[:, -1],
+                prime_fix[:, -1].value,
             )
             if ndim > 2:
                 np.testing.assert_equal(
-                    np_fix[:, :-1, ...], prime_fix[:, :-1, ...].value,
+                    np_fix[:, :-1, ...],
+                    prime_fix[:, :-1, ...].value,
                 )
                 np.testing.assert_equal(
-                    np_fix[:, :1, ...], prime_fix[:, :1, ...].value,
+                    np_fix[:, :1, ...],
+                    prime_fix[:, :1, ...].value,
                 )
                 np.testing.assert_equal(
-                    np_fix[:, 1:, ...], prime_fix[:, 1:, ...].value,
+                    np_fix[:, 1:, ...],
+                    prime_fix[:, 1:, ...].value,
                 )
             elif ndim == 2:
                 np.testing.assert_equal(
-                    np_fix[:, :2], prime_fix[:, :-1].value,
+                    np_fix[:, :2],
+                    prime_fix[:, :-1].value,
                 )
                 np.testing.assert_equal(
-                    np_fix[:, 1:], prime_fix[:, 1:].value,
+                    np_fix[:, 1:],
+                    prime_fix[:, 1:].value,
                 )
 
     @unittest.skip
@@ -80,36 +86,32 @@ class TestPrimeTensor(unittest.TestCase):
         self.set_up_indexing()
         for np_fix, prime_fix in zip(self.np_fixtures, self.prime_fixtures):
             np.testing.assert_equal(
-                np_fix[0, ...], prime_fix[0, ...].value,
+                np_fix[0, ...],
+                prime_fix[0, ...].value,
             )
             np.testing.assert_equal(
-                np_fix[1, ...], prime_fix[1, ...].value,
+                np_fix[1, ...],
+                prime_fix[1, ...].value,
             )
             np.testing.assert_equal(
-                np_fix[..., -1], prime_fix[..., -1].value,
+                np_fix[..., -1],
+                prime_fix[..., -1].value,
             )
 
     def test_arithmetic(self) -> None:
-        prime_factory = native_factory(tf.int32, 2 ** 16)
+        prime_factory = native_factory(tf.int32, 2**16)
 
-        x = prime_factory.tensor(tf.constant([2 ** 16, 2 ** 16 + 1]))
-        y = prime_factory.tensor(tf.constant([2 ** 16 + 2, 2]))
+        x = prime_factory.tensor(tf.constant([2**16, 2**16 + 1]))
+        y = prime_factory.tensor(tf.constant([2**16 + 2, 2]))
 
-        with tf.Session() as sess:
-            z = (x * y).value
-            z0 = sess.run(z)
+        z0 = (x * y).value
+        np.testing.assert_array_equal(z0, np.array([0, 2]))
 
-            np.testing.assert_array_equal(z0, np.array([0, 2]))
+        z1 = (x + y).value
+        np.testing.assert_array_equal(z1, np.array([2, 3]))
 
-            z = (x + y).value
-            z1 = sess.run(z)
-
-            np.testing.assert_array_equal(z1, np.array([2, 3]))
-
-            z = (x - y).value
-            z2 = sess.run(z)
-
-            np.testing.assert_array_equal(z2, np.array([65534, 65535]))
+        z2 = (x - y).value
+        np.testing.assert_array_equal(z2, np.array([65534, 65535]))
 
     def test_binarize(self) -> None:
         prime_factory = native_factory(tf.int32, 1001)
@@ -130,9 +132,7 @@ class TestPrimeTensor(unittest.TestCase):
             ]
         ).reshape([3, 10])
 
-        with tf.Session() as sess:
-            actual = sess.run(y.value)
-
+        actual = y.value
         np.testing.assert_array_equal(actual, expected)
 
 
